refactor: User의 Name 원시값 포장
